requirements.txt
python-telegram-bot==21.0
httpx==0.28.1
Pillow==12.0.0
python-dotenv==1.2.1
requests
beautifulsoup4